+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Notable changes to this project.
 
 ## dev
 - support passing `round_num` to `list_datasets`, to get available files from past rounds
+- `download_dataset` no longer requires a dstination path, it defaults to the source file name
 
 ## [2.8.1] - 2021-09-08
 -  Add version arg to upload_predictions (#59)